requirements.txt
cairocffi>=1.2.0
CairoSVG>=2.5.1
cffi>=1.14.4
cssselect2>=0.4.1
defusedxml>=0.6.0
Pillow>=8.1.0
pycparser>=2.20
tinycss2>=1.1.0
webencodings>=0.5.1
